Parallel: redesign dispatch, slot StateDB reuse...
** dispatch enhance
  remove atomic idle, curExec... replace by pendingExec for slot.

** slot StateDB reuse
It will try to reuse the latest merged slotDB in the same slot.
If reuse failed(conflict), it will try to update to the latest world state and redo.
The reuse SlotDB will the same BaseTxIndex, since its world state was sync when it was created based on that txIndex
Conflict check can skip current slot now.
it is more aggressive to reuse SlotDB for idle dispatch
not only pending Txs but also the idle dispatched Txs try to reuse SlotDB now.

** others
state change no needs to store value